fix(s7comm): address the final review
- The TTL's "parity-safe, only ever fires late" claim was false: the TTL
  counts processing time and each key's last-write time is restored with a
  checkpoint, so after an outage longer than the TTL every S7 connection
  restarts from empty state, silently (and so does a keepalive-held idle
  connection). The operator comment, main(), .env.example, the spec's
  section 7 and CLAUDE.md now say so and how to choose the TTL; a test pins
  the behaviour (a 61-minute outage expires the connection). The one-hour
  default is unchanged.
- The decimal pattern backtracked quadratically: one malformed record of
  100,000 digits took minutes. It is now unambiguous (the fraction needs its
  '.') and accepts exactly the same forms; the test runs in well under 2 s.
- sameFunctionRun and sameDirectionRun are long: an int wraps negative on a
  connection polled with one function for months (~248 days at 100/s).
- Deployment rule recorded (CLAUDE.md, .env.example): every protocol's
  topics must exist before either job starts, or the whole job, all four
  protocols, crash-loops.

+++ modules/domain/src/main/java/io/netsecml/platform/domain/feature/S7commConnectionState.java
@@ -48,11 +48,13 @@ public final class S7commConnectionState {
     // Scalars: the previous request function (both builders keep one; they
     // are updated under the same condition, so one field serves both), the
     // runs, the previous direction and ROSCTR, and the last timestamp (for
-    // the out-of-order flag only).
+    // the out-of-order flag only). The runs are long because upstream's are
+    // unbounded Python ints: a connection polled with one function for months
+    // would wrap an int negative.
     private Integer lastRequestFunction;
-    private int sameFunctionRun;
+    private long sameFunctionRun;
     private Boolean previousWasRequest;
-    private int sameDirectionRun;
+    private long sameDirectionRun;
     private Integer lastRosctr;
     private Double lastTs;
 
@@ -171,12 +173,12 @@ public final class S7commConnectionState {
     }
 
     // s7_same_function_run_length.
-    public int sameFunctionRunLength() {
+    public long sameFunctionRunLength() {
         return sameFunctionRun;
     }
 
     // s7_same_direction_run_length.
-    public int sameDirectionRunLength() {
+    public long sameDirectionRunLength() {
         return sameDirectionRun;
     }
